Add rose-conv tool.
This is to convert ROSE file formats to those for easier editing, for now
this only supports generating a JSON schema for QSD, and converting QSD
to yaml or json.

By using the JSON schema you can edit a .yaml file converted from QSD with
autocomplete and syntax checking in your favourite editor which supports
schemas. I have tested it with the redhat yaml extension in vscode
https://marketplace.visualstudio.com/items?itemName=redhat.vscode-yaml